requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
streamlit-authenticator>=0.2.3
PyYAML>=6.0
bcrypt>=4.0.0
streamlit_qrcode_scanner>=0.1.2
openpyxl>=3.1.0
python-dotenv>=1.0.0
openai>=1.0.0
